getLongPathValueByIndex 도입하여, 상세페이지에 적절한 게시물 노출

+++ src/main/java/com/sbs/jsp/board/boundedContext/article/repository/ArticleRepository.java
@@ -46,7 +46,7 @@ public class ArticleRepository {
     return id;
   }
 
-  public Article findById(int id) {
+  public Article findById(long id) {
     return articleList.stream().filter(
             article -> article.getId() == id
         )
